LayoutTests/fast/spatial-navigation/snav-two-elements-one-line-expected.txt

AAAAAAAAAA AAAAAAAAAA AA AA AAAAAAAAAA AAAAAAAAAA


AAAAAAAAAA AAAAAAAAAA AA B AA AAAAAAAAAA AAAAAAAAAA
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test is testing that we can handle 2 inline elements in the same line.
